Refactored some functionality into new files.
